refactor(llc): move the channel extensions into their own files
