+++ packages/stream_chat/test/src/client/channel_test.dart
@@ -8915,594 +8915,6 @@ void main() {
     });
   });
 
-  group('ChannelReadHelper', () {
-    const channelId = 'test-channel-id';
-    const channelType = 'test-channel-type';
-    late final client = MockStreamChatClient();
-
-    // A date in the distant past (Unix epoch), useful for representing old dates
-    final distantPast = DateTime.fromMillisecondsSinceEpoch(0, isUtc: true);
-
-    setUpAll(() {
-      // detached loggers
-      when(() => client.detachedLogger(any())).thenAnswer((invocation) {
-        final name = invocation.positionalArguments.first;
-        return _createLogger(name);
-      });
-
-      final retryPolicy = RetryPolicy(
-        shouldRetry: (_, __, ___) => false,
-        delayFactor: Duration.zero,
-      );
-      when(() => client.retryPolicy).thenReturn(retryPolicy);
-
-      // fake clientState
-      final clientState = FakeClientState();
-      when(() => client.state).thenReturn(clientState);
-
-      // client logger
-      when(() => client.logger).thenReturn(_createLogger('mock-client-logger'));
-    });
-
-    test('userReadOf should return read for specific user', () {
-      final now = DateTime.now();
-      final user1 = User(id: 'user-1', name: 'User 1');
-      final user2 = User(id: 'user-2', name: 'User 2');
-
-      final reads = [
-        Read(user: user1, lastRead: now),
-        Read(user: user2, lastRead: now.add(const Duration(minutes: 1))),
-      ];
-
-      final channelState = _generateChannelState(channelId, channelType);
-      final channel = Channel.fromState(client, channelState);
-      addTearDown(channel.dispose);
-
-      channel.state!.updateChannelState(
-        ChannelState(channel: channelState.channel, read: reads),
-      );
-
-      final user1Read = channel.state!.userReadOf(userId: 'user-1');
-      expect(user1Read, isNotNull);
-      expect(user1Read!.user.id, 'user-1');
-      expect(user1Read.lastRead, now);
-
-      final user2Read = channel.state!.userReadOf(userId: 'user-2');
-      expect(user2Read, isNotNull);
-      expect(user2Read!.user.id, 'user-2');
-
-      final nonExistentRead = channel.state!.userReadOf(userId: 'user-3');
-      expect(nonExistentRead, isNull);
-    });
-
-    test('userReadOf should return null when userId is null', () {
-      final channelState = _generateChannelState(channelId, channelType);
-      final channel = Channel.fromState(client, channelState);
-      addTearDown(channel.dispose);
-
-      final read = channel.state!.userReadOf(userId: null);
-      expect(read, isNull);
-    });
-
-    test(
-      'userReadStreamOf should emit read updates for specific user',
-      () async {
-        final now = DateTime.now();
-        final user1 = User(id: 'user-1', name: 'User 1');
-
-        final channelState = _generateChannelState(channelId, channelType);
-        final channel = Channel.fromState(client, channelState);
-        addTearDown(channel.dispose);
-
-        final readStream = channel.state!.userReadStreamOf(userId: 'user-1');
-
-        expectLater(
-          readStream,
-          emitsInOrder([
-            isNull, // initial state
-            isA<Read>().having((r) => r.user.id, 'userId', 'user-1'),
-          ]),
-        );
-
-        // Update with read
-        channel.state!.updateChannelState(
-          ChannelState(
-            channel: channelState.channel,
-            read: [Read(user: user1, lastRead: now)],
-          ),
-        );
-      },
-    );
-
-    test('readsOf should return reads that have marked message as read', () {
-      final now = DateTime.now();
-      final sender = User(id: 'sender-id', name: 'Sender');
-      final user1 = User(id: 'user-1', name: 'User 1');
-      final user2 = User(id: 'user-2', name: 'User 2');
-      final user3 = User(id: 'user-3', name: 'User 3');
-
-      final message = Message(
-        id: 'msg-1',
-        text: 'Test message',
-        user: sender,
-        createdAt: now,
-      );
-
-      final reads = [
-        // user1 has read the message
-        Read(user: user1, lastRead: now.add(const Duration(seconds: 1))),
-        // user2 has not read the message yet
-        Read(user: user2, lastRead: distantPast),
-        // user3 has read the message
-        Read(user: user3, lastRead: now.add(const Duration(seconds: 2))),
-        // sender should be excluded
-        Read(user: sender, lastRead: now.add(const Duration(seconds: 10))),
-      ];
-
-      final channelState = _generateChannelState(channelId, channelType);
-      final channel = Channel.fromState(client, channelState);
-      addTearDown(channel.dispose);
-
-      channel.state!.updateChannelState(
-        ChannelState(channel: channelState.channel, read: reads),
-      );
-
-      final messageReads = channel.state!.readsOf(message: message);
-      expect(messageReads.length, 2);
-      expect(messageReads.map((r) => r.user.id), containsAll(['user-1', 'user-3']));
-      expect(messageReads.map((r) => r.user.id), isNot(contains('user-2')));
-      expect(messageReads.map((r) => r.user.id), isNot(contains('sender-id')));
-    });
-
-    test('readsOfStream should emit read updates for a message', () async {
-      final now = DateTime.now();
-      final sender = User(id: 'sender-id', name: 'Sender');
-      final user1 = User(id: 'user-1', name: 'User 1');
-
-      final message = Message(
-        id: 'msg-1',
-        text: 'Test message',
-        user: sender,
-        createdAt: now,
-      );
-
-      final channelState = _generateChannelState(channelId, channelType);
-      final channel = Channel.fromState(client, channelState);
-      addTearDown(channel.dispose);
-
-      final readsStream = channel.state!.readsOfStream(message: message);
-
-      expectLater(
-        readsStream,
-        emitsInOrder([
-          isEmpty, // initial state
-          hasLength(1), // after adding read
-        ]),
-      );
-
-      // Update with read
-      channel.state!.updateChannelState(
-        ChannelState(
-          channel: channelState.channel,
-          read: [Read(user: user1, lastRead: now.add(const Duration(seconds: 1)))],
-        ),
-      );
-    });
-
-    test('deliveriesOf should return reads that have delivered the message', () {
-      final now = DateTime.now();
-      final sender = User(id: 'sender-id', name: 'Sender');
-      final user1 = User(id: 'user-1', name: 'User 1');
-      final user2 = User(id: 'user-2', name: 'User 2');
-      final user3 = User(id: 'user-3', name: 'User 3');
-      final user4 = User(id: 'user-4', name: 'User 4');
-
-      final message = Message(
-        id: 'msg-1',
-        text: 'Test message',
-        user: sender,
-        createdAt: now,
-      );
-
-      final reads = [
-        // user1 has delivered the message
-        Read(
-          user: user1,
-          lastRead: distantPast,
-          lastDeliveredAt: now.add(const Duration(seconds: 1)),
-        ),
-        // user2 has not delivered the message yet (lastDeliveredAt is before message)
-        Read(
-          user: user2,
-          lastRead: distantPast,
-          lastDeliveredAt: distantPast,
-        ),
-        // user3 has no lastDeliveredAt
-        Read(
-          user: user3,
-          lastRead: distantPast,
-        ),
-        // user4 has read the message (implicitly delivered)
-        Read(
-          user: user4,
-          lastRead: now.add(const Duration(seconds: 1)),
-        ),
-        // sender should be excluded
-        Read(
-          user: sender,
-          lastRead: now.add(const Duration(seconds: 10)),
-          lastDeliveredAt: now.add(const Duration(seconds: 10)),
-        ),
-      ];
-
-      final channelState = _generateChannelState(channelId, channelType);
-      final channel = Channel.fromState(client, channelState);
-      addTearDown(channel.dispose);
-
-      channel.state!.updateChannelState(
-        ChannelState(channel: channelState.channel, read: reads),
-      );
-
-      final deliveries = channel.state!.deliveriesOf(message: message);
-      expect(deliveries.length, 2);
-      expect(deliveries.map((r) => r.user.id), containsAll(['user-1', 'user-4']));
-      expect(deliveries.map((r) => r.user.id), isNot(contains('user-2')));
-      expect(deliveries.map((r) => r.user.id), isNot(contains('user-3')));
-      expect(deliveries.map((r) => r.user.id), isNot(contains('sender-id')));
-    });
-
-    test('deliveriesOfStream should emit delivery updates for a message', () async {
-      final now = DateTime.now();
-      final sender = User(id: 'sender-id', name: 'Sender');
-      final user1 = User(id: 'user-1', name: 'User 1');
-
-      final message = Message(
-        id: 'msg-1',
-        text: 'Test message',
-        user: sender,
-        createdAt: now,
-      );
-
-      final channelState = _generateChannelState(channelId, channelType);
-      final channel = Channel.fromState(client, channelState);
-      addTearDown(channel.dispose);
-
-      final deliveriesStream = channel.state!.deliveriesOfStream(message: message);
-
-      expectLater(
-        deliveriesStream,
-        emitsInOrder([
-          isEmpty, // initial state
-          hasLength(1), // after adding delivery
-        ]),
-      );
-
-      // Update with delivery
-      channel.state!.updateChannelState(
-        ChannelState(
-          channel: channelState.channel,
-          read: [
-            Read(
-              user: user1,
-              lastRead: distantPast,
-              lastDeliveredAt: now.add(const Duration(seconds: 1)),
-            ),
-          ],
-        ),
-      );
-    });
-  });
-
-  group('ChannelCapabilityCheck', () {
-    const channelId = 'test-channel-id';
-    const channelType = 'test-channel-type';
-    late final client = MockStreamChatClient();
-
-    setUpAll(() {
-      // detached loggers
-      when(() => client.detachedLogger(any())).thenAnswer((invocation) {
-        final name = invocation.positionalArguments.first;
-        return _createLogger(name);
-      });
-
-      final retryPolicy = RetryPolicy(
-        shouldRetry: (_, __, ___) => false,
-        delayFactor: Duration.zero,
-      );
-      when(() => client.retryPolicy).thenReturn(retryPolicy);
-
-      // fake clientState
-      final clientState = FakeClientState();
-      when(() => client.state).thenReturn(clientState);
-
-      // client logger
-      when(() => client.logger).thenReturn(_createLogger('mock-client-logger'));
-    });
-
-    /// Parameterized test for channel capability extension properties
-    void testCapability(
-      String capabilityName,
-      ChannelCapability capability,
-      bool Function(Channel) getterMethod,
-    ) {
-      test('can$capabilityName returns false when capability is absent', () {
-        final channelState = _generateChannelState(channelId, channelType);
-        final channel = Channel.fromState(client, channelState);
-        expect(getterMethod(channel), false);
-      });
-
-      test('can$capabilityName returns true when capability is present', () {
-        final channelState = _generateChannelState(
-          channelId,
-          channelType,
-          ownCapabilities: [capability],
-        );
-        final channel = Channel.fromState(client, channelState);
-        expect(getterMethod(channel), true);
-      });
-    }
-
-    // Test all channel capabilities using the parameterized function
-    testCapability(
-      'SendMessage',
-      ChannelCapability.sendMessage,
-      (channel) => channel.canSendMessage,
-    );
-
-    testCapability(
-      'SendReply',
-      ChannelCapability.sendReply,
-      (channel) => channel.canSendReply,
-    );
-
-    testCapability(
-      'SendRestrictedVisibilityMessage',
-      ChannelCapability.sendRestrictedVisibilityMessage,
-      (channel) => channel.canSendRestrictedVisibilityMessage,
-    );
-
-    testCapability(
-      'SendReaction',
-      ChannelCapability.sendReaction,
-      (channel) => channel.canSendReaction,
-    );
-
-    testCapability(
-      'SendLinks',
-      ChannelCapability.sendLinks,
-      (channel) => channel.canSendLinks,
-    );
-
-    testCapability(
-      'CreateAttachment',
-      ChannelCapability.createAttachment,
-      (channel) => channel.canCreateAttachment,
-    );
-
-    testCapability(
-      'FreezeChannel',
-      ChannelCapability.freezeChannel,
-      (channel) => channel.canFreezeChannel,
-    );
-
-    testCapability(
-      'SetChannelCooldown',
-      ChannelCapability.setChannelCooldown,
-      (channel) => channel.canSetChannelCooldown,
-    );
-
-    testCapability(
-      'LeaveChannel',
-      ChannelCapability.leaveChannel,
-      (channel) => channel.canLeaveChannel,
-    );
-
-    testCapability(
-      'JoinChannel',
-      ChannelCapability.joinChannel,
-      (channel) => channel.canJoinChannel,
-    );
-
-    testCapability(
-      'PinMessage',
-      ChannelCapability.pinMessage,
-      (channel) => channel.canPinMessage,
-    );
-
-    testCapability(
-      'DeleteAnyMessage',
-      ChannelCapability.deleteAnyMessage,
-      (channel) => channel.canDeleteAnyMessage,
-    );
-
-    testCapability(
-      'DeleteOwnMessage',
-      ChannelCapability.deleteOwnMessage,
-      (channel) => channel.canDeleteOwnMessage,
-    );
-
-    testCapability(
-      'UpdateAnyMessage',
-      ChannelCapability.updateAnyMessage,
-      (channel) => channel.canUpdateAnyMessage,
-    );
-
-    testCapability(
-      'UpdateOwnMessage',
-      ChannelCapability.updateOwnMessage,
-      (channel) => channel.canUpdateOwnMessage,
-    );
-
-    testCapability(
-      'SearchMessages',
-      ChannelCapability.searchMessages,
-      (channel) => channel.canSearchMessages,
-    );
-
-    testCapability(
-      'SendTypingEvents',
-      ChannelCapability.sendTypingEvents,
-      (channel) => channel.canSendTypingEvents,
-    );
-
-    testCapability(
-      'UploadFile',
-      ChannelCapability.uploadFile,
-      (channel) => channel.canUploadFile,
-    );
-
-    testCapability(
-      'DeleteChannel',
-      ChannelCapability.deleteChannel,
-      (channel) => channel.canDeleteChannel,
-    );
-
-    testCapability(
-      'UpdateChannel',
-      ChannelCapability.updateChannel,
-      (channel) => channel.canUpdateChannel,
-    );
-
-    testCapability(
-      'UpdateChannelMembers',
-      ChannelCapability.updateChannelMembers,
-      (channel) => channel.canUpdateChannelMembers,
-    );
-
-    testCapability(
-      'UpdateThread',
-      ChannelCapability.updateThread,
-      (channel) => channel.canUpdateThread,
-    );
-
-    testCapability(
-      'QuoteMessage',
-      ChannelCapability.quoteMessage,
-      (channel) => channel.canQuoteMessage,
-    );
-
-    testCapability(
-      'BanChannelMembers',
-      ChannelCapability.banChannelMembers,
-      (channel) => channel.canBanChannelMembers,
-    );
-
-    testCapability(
-      'FlagMessage',
-      ChannelCapability.flagMessage,
-      (channel) => channel.canFlagMessage,
-    );
-
-    testCapability(
-      'MuteChannel',
-      ChannelCapability.muteChannel,
-      (channel) => channel.canMuteChannel,
-    );
-
-    testCapability(
-      'SendCustomEvents',
-      ChannelCapability.sendCustomEvents,
-      (channel) => channel.canSendCustomEvents,
-    );
-
-    testCapability(
-      'ReceiveReadEvents',
-      ChannelCapability.readEvents,
-      (channel) => channel.canReceiveReadEvents,
-    );
-
-    testCapability(
-      'ReceiveConnectEvents',
-      ChannelCapability.connectEvents,
-      (channel) => channel.canReceiveConnectEvents,
-    );
-
-    testCapability(
-      'UseTypingEvents',
-      ChannelCapability.typingEvents,
-      (channel) => channel.canUseTypingEvents,
-    );
-
-    testCapability(
-      'InSlowMode',
-      ChannelCapability.slowMode,
-      (channel) => channel.isInSlowMode,
-    );
-
-    testCapability(
-      'SkipSlowMode',
-      ChannelCapability.skipSlowMode,
-      (channel) => channel.canSkipSlowMode,
-    );
-
-    testCapability(
-      'SendPoll',
-      ChannelCapability.sendPoll,
-      (channel) => channel.canSendPoll,
-    );
-
-    testCapability(
-      'CastPollVote',
-      ChannelCapability.castPollVote,
-      (channel) => channel.canCastPollVote,
-    );
-
-    testCapability(
-      'QueryPollVotes',
-      ChannelCapability.queryPollVotes,
-      (channel) => channel.canQueryPollVotes,
-    );
-
-    testCapability(
-      'ShareLocation',
-      ChannelCapability.shareLocation,
-      (channel) => channel.canShareLocation,
-    );
-
-    testCapability(
-      'NotifyChannel',
-      ChannelCapability.notifyChannel,
-      (channel) => channel.canNotifyChannel,
-    );
-
-    testCapability(
-      'NotifyHere',
-      ChannelCapability.notifyHere,
-      (channel) => channel.canNotifyHere,
-    );
-
-    testCapability(
-      'NotifyRole',
-      ChannelCapability.notifyRole,
-      (channel) => channel.canNotifyRole,
-    );
-
-    testCapability(
-      'NotifyGroup',
-      ChannelCapability.notifyGroup,
-      (channel) => channel.canNotifyGroup,
-    );
-
-    test('returns correct values with multiple capabilities', () {
-      final channelState = _generateChannelState(
-        channelId,
-        channelType,
-        ownCapabilities: [
-          ChannelCapability.sendMessage,
-          ChannelCapability.sendReply,
-          ChannelCapability.deleteOwnMessage,
-        ],
-      );
-
-      final channel = Channel.fromState(client, channelState);
-      expect(channel.canSendMessage, true);
-      expect(channel.canSendReply, true);
-      expect(channel.canDeleteOwnMessage, true);
-      expect(channel.canDeleteAnyMessage, false);
-      expect(channel.canUpdateChannel, false);
-    });
-  });
-
   group('Channel State Validation and Cooldown', () {
     late final client = MockStreamChatClient();
     const channelId = 'test-channel-id';
